requirements.txt
playwright==1.42.0
python-dotenv==1.0.1
pydantic==2.6.3
loguru==0.7.2
pytest==8.0.2
pytest-asyncio==0.23.5
python-dateutil==2.8.2 